requirements.txt
vanna[gemini,sqlite,openai,ollama]>=2.0.0
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
plotly>=6.0.0
pandas>=2.2.0
python-dotenv>=1.0.1
google-genai>=1.30.0
openai>=1.99.0
ollama>=0.5.0
